Fix aws-sdk-go test behavior
Added missing endpoint entry in the case of secure endpoint.

Added bucket creation and deletion code, as the error code
is different between AWS S3 and Minio server, if test is run
with a non-existent bucket.

Added argument cleanup function

Added name of the test under test to the "name" field in the log.

Fixes #228